requirements.txt
mysql-connector-python
sqlalchemy pymysql
pymysql
plotly
python-dotenv
pandas